requirements.txt
torch>=1.13.0
torchvision
transformers>=4.30.0
tqdm
pandas
Pillow
numpy
matplotlib